feat(cli): neatly handle invalid directory
Raise a user friendly CLick error
if user passes a directory that doesn't exist to the read command.

Closes #5

+++ heave/cli.py
@@ -1,5 +1,5 @@
 """Heave CLI."""
-
+from pathlib import Path
 
 import click
 from sqlalchemy import create_engine
@@ -102,7 +102,7 @@ def cli(
 
 
 @cli.command()
-@click.argument("path", type=click.Path(exists=True))
+@click.argument("path", type=click.Path(exists=True, readable=True, path_type=Path))
 @click.option("-t", "--table", required=True, help="Table to insert into.")
 @click.option("-s", "--schema", help="Table schema name.")
 @click.option(
@@ -112,7 +112,7 @@ def cli(
     help="Handle conflict errors.",
 )
 @click.pass_obj
-def insert(obj, path: str, table: str, schema: str | None, on_conflict: str | None):
+def insert(obj, path: Path, table: str, schema: str | None, on_conflict: str | None):
     """Insert data from a file into a table."""
     data = file.read_csv(path)
     sql_table = sql.reflect_table(obj, table, schema)
@@ -121,12 +121,14 @@ def insert(obj, path: str, table: str, schema: str | None, on_conflict: str | No
 
 
 @cli.command()
-@click.argument("path", type=click.Path(exists=False))
+@click.argument("path", type=click.Path(exists=False, writable=True, path_type=Path))
 @click.option("-t", "--table", required=True, help="Table to read.")
 @click.option("-s", "--schema", help="Table schema name.")
 @click.pass_obj
-def read(obj, path: str, table: str, schema: str | None):
+def read(obj, path: Path, table: str, schema: str | None):
     """Read data from a table and write it to a file."""
+    if not (d := path.parent).exists():
+        raise click.ClickException(f"No such directory: '{d}'")
     sql_table = sql.reflect_table(obj, table, schema)
     data = sql.read(obj, sql_table)
     file.write_csv(data, path)
